docker_requirements.txt
beautifulsoup4==4.13.4
playwright==1.51.0
python-dotenv==1.1.0
pytz
tqdm